+++ src/pages/recipe/RecipePage.svelte
@@ -1,11 +1,10 @@
 <script>
-  import Button, { Label } from '@smui/button';
-  import { PrimaryAction } from '@smui/card';
-  import Dialog, { Actions } from '@smui/dialog';
+  import Button from '@smui/button';
   import { onMount } from 'svelte';
   import InfiniteScroll from 'svelte-infinite-scroll';
   import { navigate } from 'svelte-routing';
   import RecipeProcesses from '../../components/recipe/RecipeProcesses.svelte';
+  import RecipeUpdateComponent from '../../components/recipe/RecipeUpdateComponent.svelte';
   import RecipeFoodAllCard from '../../components/recipefood/RecipeFoodAllCard.svelte';
   import ReviewCard from '../../components/review/ReviewCard.svelte';
   import { REVIEW_SIZE } from '../../lib/const/pagination-const';
@@ -18,7 +17,7 @@
   import { auth, isLogin } from '../../store/user';
 
   export let recipeId;
-  let open = false;
+
   let totalPrice;
   let recipe = {};
 
@@ -228,68 +227,6 @@
         alert('레시피 삭제에 실패했습니다.');
       });
   };
-  const MAX_FILE_SIZE = 2 * 1024 * 1024;
-
-  const handleSelectRecipeImage = (e) => {
-    const fileInput = e.target;
-    const file = e.target.files[0];
-
-    if (file && file.size > MAX_FILE_SIZE) {
-      alert('파일 크기가 너무 큽니다. 2MB 이하의 파일을 선택해 주세요.');
-      fileInput.value = '';
-      return;
-    }
-
-    recipeUpdateImage = file;
-  };
-
-  const updateRecipe = () => {
-    let formData = new FormData();
-    formData.append(
-      'recipeUpdateRequestDto',
-      new Blob([JSON.stringify(recipeUpdateRequestDto)], {
-        type: 'application/json',
-      }),
-    );
-    if (recipeUpdateImage == null) {
-      formData.append('recipeUpdateImage', new Blob());
-    } else {
-      let recipeUpdateImageType;
-      if (recipeUpdateImage.type == 'image/png') {
-        recipeUpdateImageType = 'image/png';
-      } else {
-        recipeUpdateImageType = 'image/jpeg';
-      }
-      formData.append(
-        'recipeUpdateImage',
-        new Blob([recipeUpdateImage], { type: recipeUpdateImageType }),
-      );
-    }
-    fetch(HOST + `/api/v1/recipes/${recipeId}`, {
-      method: 'PUT',
-      headers: {
-        Authorization: $auth.Authorization,
-      },
-      body: formData,
-    })
-      .then((response) => {
-        if (response.status >= 400 && response.status < 600) {
-          throw response;
-        }
-        alert('레시피가 성공적으로 업데이트되었습니다.');
-        recipe = { ...recipeUpdateRequestDto };
-      })
-      .catch((error) => {
-        alert(recipeUpdateRequestDto.title);
-      });
-  };
-  const handleCloseDialog = (e) => {
-    switch (e.detail.action) {
-      case 'save':
-        updateRecipe();
-        break;
-    }
-  };
 </script>
 
 <link
@@ -312,96 +249,18 @@
       {/if}
     </button>
 
-    {#if $isLogin}
-      {#if $auth.nickname == recipe.writer}
-        <button
-          class="delete-button"
-          style="color: black;"
-          on:click={deleteRecipe}
-        >
-          <span class="material-icons">delete</span>
-        </button>
-      {/if}
-    {:else}{/if}
-  </div>
-  {#if $isLogin}
-    {#if $auth.nickname == recipe.writer}
-      <PrimaryAction on:click={() => (open = !open)}>
-        <h1 class="recipe-title">{recipe.title}</h1>
-        <h3 class="recipe-intro">
-          {recipe.intro}
-        </h3>
-        <span class="recipe-serving">
-          {recipe.serving}인분
-        </span>
-        <hr class="hr-100" />
-      </PrimaryAction>
-      <Dialog
-        class="custom-dialog"
-        bind:open
-        aria-labelledby="simple-title"
-        aria-describedby="simple-content"
-        on:SMUIDialog:closed={handleCloseDialog}
+    {#if $isLogin && $auth.id == recipe.writerId}
+      <button
+        class="delete-button"
+        style="color: black;"
+        on:click={deleteRecipe}
       >
-        <div class="recipe-form">
-          <div class="form-group">
-            <label for="recipe-title">레시피 제목</label>
-            <input
-              required
-              type="text"
-              id="recipe-title"
-              bind:value={recipeUpdateRequestDto.title}
-            />
-          </div>
-          <div class="form-group">
-            <label for="recipe-title">레시피 대표 사진</label>
-            <input
-              type="file"
-              accept=".jpg, .jpeg, .png"
-              on:change={(e) => handleSelectRecipeImage(e)}
-            />
-          </div>
-          <div class="form-group">
-            <label for="intro">레시피 소개</label>
-            <textarea id="intro" bind:value={recipeUpdateRequestDto.intro}
-            ></textarea>
-          </div>
-
-          <div class="form-group">
-            <label for="serving">인원</label>
-            <select id="serving" bind:value={recipeUpdateRequestDto.serving}>
-              <option value="1">1인분</option>
-              <option value="2">2인분</option>
-              <option value="3">3인분</option>
-              <option value="4">4인분</option>
-              <option value="5">5인분</option>
-              <option value="6">6인분 이상</option>
-            </select>
-          </div>
-        </div>
-        <Actions>
-          <div class="btn-container">
-            <span>
-              <Button action="save">
-                <Label>저장</Label>
-              </Button>
-              <Button action="cancel">
-                <Label>취소</Label>
-              </Button>
-            </span>
-          </div>
-        </Actions>
-      </Dialog>
-    {:else}
-      <h1 class="recipe-title">{recipe.title}</h1>
-      <h3 class="recipe-intro">
-        {recipe.intro}
-      </h3>
-      <span class="recipe-serving">
-        {recipe.serving}인분
-      </span>
-      <hr class="hr-100" />
+        <span class="material-icons">delete</span>
+      </button>
     {/if}
+  </div>
+  {#if $isLogin && $auth.id == recipe.writerId}
+    <RecipeUpdateComponent {recipeId} bind:recipe />
   {:else}
     <h1 class="recipe-title">{recipe.title}</h1>
     <h3 class="recipe-intro">
@@ -412,6 +271,7 @@
     </span>
     <hr class="hr-100" />
   {/if}
+  <hr class="hr-100" />
   <div class="container-flex">
     <RecipeFoodAllCard {recipeFoods} bind:totalPrice bind:selectedRecipeFoods />
     <br />
@@ -683,21 +543,5 @@
   .comment-input-container {
     flex-grow: 2;
     margin-right: 20px;
-  }
-  .recipe-form {
-    background-color: rgba(0, 0, 0, 0.6);
-    padding: 20px;
-    margin: 0 auto;
-    width: auto;
-    height: auto;
-    border-radius: 10px;
-  }
-  .custom-dialog {
-    background-color: rgba(0, 0, 0, 0.6);
-    padding: 20px;
-    margin: 0 auto;
-    width: auto;
-    height: auto;
-    border-radius: 10px;
   }
 </style>
